feat: Se actualiza la configuracion de seguridad sin secretos

+++ src/main/java/com/taxisaeropuerto/taxisAeropuerto/repository/RolRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RolRepository extends JpaRepository<Rol, Integer> {
+public interface RolRepository extends JpaRepository<Rol, Long> { // <-- ¡Cambia a Long!
     Optional<Rol> findByNombre(String nombre);
 }
